remove cyclic
